Convert the theme palettes to TypeScript
`Styles/Themes/{dark,light,index}` move to `.ts`. The palettes change only
their `module.exports = {` line into `const <name> = {` plus an
`export default`; no colour value moves. `index` swaps its two
`import * as` namespace imports for default imports and gains `ThemeName`
and `Theme`.

`Theme` is derived from the two palettes rather than written out, so a
variable added to one and not the other is a compiler error -- except for
the four that already diverge, which are listed as optional so the gap is
named instead of hidden behind an index signature. Three of the four are
read from CSS by the palette that lacks them: `--themeLightPurple` and
`--sceneBackgroundColor` are dark-only and read by `PageSidebarItem.css`
and five poster stylesheets, `--whisparrPurple` is light-only and read by
four poster stylesheets, so each resolves to nothing in the other theme.
`themeAlternatePurple` is light-only and read by nothing. Those are
palette bugs rather than conversion ones and are fixed separately.

`Styles/Variables/*` stay `.js` -- `postcss.config.js` `require`s them
from Node to feed `postcss-simple-vars`, and Node cannot require `.ts`.

The default export is anonymous because `filenames/match-exported`
otherwise wants the `Themes` directory renamed to match it.
