--- frontend/src/Styles/Themes/index.js
@@ -1,11 +1,0 @@
-import * as dark from './dark';
-import * as light from './light';
-
-const defaultDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
-const auto = defaultDark ? dark : light;
-
-export default {
-  auto,
-  light,
-  dark,
-};
